Move MovieEntryAsyncTask from the inner class of MainActivity to a separate class in the sync folder
