+++ .gitignore
@@ -9,3 +9,5 @@
 .externalNativeBuild
 /gradle.properties
 /app/build
+app/libs/
+app/src/main/res/values/api_key.xml
