+++ app/src/main/java/com/andrewclam/popularmovie/MainActivity.java
@@ -11,29 +11,23 @@
 package com.andrewclam.popularmovie;
 
 import android.content.Intent;
-import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.GridLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.Toolbar;
-import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.LinearLayout;
 import android.widget.ProgressBar;
 
+import com.andrewclam.popularmovie.sync.MovieEntryAsyncTask;
 import com.andrewclam.popularmovie.utilities.LayoutManagerUtils;
-import com.andrewclam.popularmovie.utilities.NetworkUtils;
-import com.andrewclam.popularmovie.utilities.TMDBJsonUtils;
 
-import org.json.JSONException;
 import org.parceler.Parcels;
 
-import java.io.IOException;
-import java.net.URL;
 import java.util.ArrayList;
 
 import static com.andrewclam.popularmovie.utilities.NetworkUtils.TMDB_PATH_POPULAR;
@@ -44,7 +38,7 @@ import static com.andrewclam.popularmovie.utilities.NetworkUtils.TMDB_PATH_TOP_R
  * <p>
  * MainActivity of the PopularMovie application, shows a user a default list of movies from TMDB
  * (first-page) in a recyclerView, user can select sort-value base on their preference and get a
- * different list of movie as a response
+ * different list of movie as a response.
  */
 
 public class MainActivity extends AppCompatActivity implements MovieEntryAdapter.OnMovieEntryClickListener {
@@ -209,90 +203,5 @@ public class MainActivity extends AppCompatActivity implements MovieEntryAdapter
         Intent intent = new Intent(MainActivity.this, DetailActivity.class);
         intent.putExtra(EXTRA_MOVIE_ENTRY_OBJECT, Parcels.wrap(entry));
         startActivity(intent);
-    }
-
-    /**
-     * MovieEntryAsyncTask
-     * A static implementation of the AsyncTask class to do network IO on a separate thread,
-     * (!) Lint suggests to make this class static to avoid memory leak
-     */
-
-    public static class MovieEntryAsyncTask extends AsyncTask<String, Void, ArrayList<MovieEntry>> {
-        private onMovieEntryTaskInteractionListener mListener;
-        private String mApiKey;
-
-        private MovieEntryAsyncTask setListener(onMovieEntryTaskInteractionListener mListener) {
-            this.mListener = mListener;
-            return this;
-        }
-
-        private MovieEntryAsyncTask setApiKey(String mApiKey) {
-            this.mApiKey = mApiKey;
-            return this;
-        }
-
-        @Override
-        protected void onPreExecute() {
-            super.onPreExecute();
-            mListener.onPreExecute();
-        }
-
-        @Override
-        protected ArrayList<MovieEntry> doInBackground(String... strings) {
-            // Get the sortByValue from the strings input
-            String sortByValue = strings[0];
-
-            // Init a arrayList to store the parsed movie entries
-            ArrayList<MovieEntry> entries;
-
-            try {
-                // Check for null error, sortByValue should not be null
-                if (sortByValue == null) return null;
-
-                // Get the url required by the network util
-                URL url = NetworkUtils.buildUrl(sortByValue, mApiKey);
-
-                // Check for null url
-                if (url == null) return null;
-
-                // Get httpResponse using the url
-                String jsonResponse = NetworkUtils.getResponseFromHttpUrl(url);
-
-                // Check for null response
-                if (jsonResponse == null) return null;
-
-                // Parse the jsonResponse using the JsonUtils
-                entries = TMDBJsonUtils.getMovieDataFromJson(jsonResponse);
-
-            } catch (IOException e) {
-                Log.e(TAG, "MovieEntryAsyncTask - doInBackground - IO Error occurred while getting the jsonResponse from the url");
-                e.printStackTrace();
-                return null;
-            } catch (JSONException e) {
-                Log.e(TAG, "MovieEntryAsyncTask - doInBackground - JSONException occurred while parsing the jsonResponse into model class");
-                e.printStackTrace();
-                return null;
-            }
-
-            // return the entries, may be empty or null
-            // handles in the callback onPostExecute
-            return entries;
-        }
-
-        @Override
-        protected void onPostExecute(ArrayList<MovieEntry> entries) {
-            super.onPostExecute(entries);
-            mListener.onPostExecute(entries);
-        }
-
-        /**
-         * Interface for callback to the listener at stages where UI change is required
-         * in preExecute and postExecute.
-         */
-        private interface onMovieEntryTaskInteractionListener {
-            void onPreExecute();
-
-            void onPostExecute(ArrayList<MovieEntry> entries);
-        }
     }
 }
